pyproject.toml
[tool.poetry]
name = "mutual-funds-insight"
version = "0.1.0"
description = "Mutual fund analytics web application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
plotly = "^5.24.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
fastapi = "^0.112.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pandas-ta = "^0.3.14b0"
pydantic = "^2.6.0"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
httpx = "^0.27.0"
openai = "^1.40.0"
gspread = "^6.0.0"
google-auth = "^2.32.0"
google-api-python-client = "^2.143.0"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^24.4.0"
isort = "^5.13.0"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
